[NO_REVIEW] Merge r70913 - "Fix the KratiStorageEngine code that builds an entries iterator to properly handle key collisions.".
The fix part of this update had already been submitted back to Voldemort and therefore is not actually included in this merge.

This merge still moves the Krati code from the contrib dir to the main code dirs, but,to make future merges easier it does not replace BDB with 
Krati in any of the existing tests.  The code move does trigger the execution of the Krati specific storage engine tests during normal
junit testing..  

A bug was found in the base Voldemort storage engine test structure.  Tests inherited by a parametrized storage engine test class would not 
be run if the test in the super class was not annotated with @Test.  To address this issue the @Test annotation was added to several tests.  

This change triggered several failures for the RocksDB storage engine (for test that previously were not actually getting executed).  The key
iteration with serialization test case failures may be important for us.  More investigation of them may be required if we decide to try
RocksDB.

The storage engine test I provided to Voldemort to demonstrate the error the original fix was intended to address was running very slowly for
Krati (I had not seen this in hte base Voldemort code).  To address this I changed how the Krati storage engine used to run the junit test is 
initialized.  I set persistOnWrite to false (the default value we use in production).